pyproject.toml
[project]
name = "stockist-dashboard"
version = "0.1.0"
description = "Super Stockist Business Dashboard"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.42.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "gspread>=6.1.0",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
